Update pipeline; fix issues with tests
